misc/HtmlTestSuite/results/test-0002.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><link href="mailto:emilio@crisal.io" rel="author" title="Emilio Cobos Ãlvarez"/>
<link href="https://mozilla.org" rel="author" title="Mozilla"/>
<link href="https://bugzilla.mozilla.org/show_bug.cgi?id=1555949" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
</head><body><div id="host"/>
<script>
  let svgScript = document.createElementNS("http://www.w3.org/2000/svg", "script");
  svgScript.innerHTML = "window.scriptRan = true";
  document.getElementById("host").attachShadow({ mode: "open" }).appendChild(svgScript);
  test(function() {
    assert_true(window.scriptRan);
  }, "svg:script runs correctly in a shadow tree");
</script>
</body></html>
